Refactoring and hook "onoverride" added
