+++ main-core/Mandarine.ns.ts
@@ -340,29 +340,6 @@ export namespace Mandarine {
             // Initialize resource handlers
             componentRegistry.get(MandarineCore.NativeComponents.WebMVCConfigurer).addResourceHandlers();
         };
-
-        /**
-        * Overrides the current session container
-        */
-        export function changeSessionContainer(newSessionContainer: MandarineSecurity.Sessions.SessionContainer): void {
-            let defaultContainer = Defaults.MandarineDefaultSessionContainer();
-            let mandarineGlobal: Mandarine.Global.MandarineGlobalInterface  = Mandarine.Global.getMandarineGlobal();
-            
-            mandarineGlobal.mandarineSessionContainer = <MandarineSecurity.Sessions.SessionContainer>{
-                cookie: {
-                    path: (newSessionContainer.cookie && newSessionContainer.cookie.path) ? newSessionContainer.cookie.path : defaultContainer.cookie.path,
-                    maxAge: (newSessionContainer.cookie && newSessionContainer.cookie.maxAge) ? newSessionContainer.cookie.maxAge : defaultContainer.cookie.maxAge,
-                    httpOnly: (newSessionContainer.cookie && newSessionContainer.cookie.httpOnly) ? newSessionContainer.cookie.httpOnly : defaultContainer.cookie.httpOnly,
-                    secure: (newSessionContainer.cookie && newSessionContainer.cookie.secure) ? newSessionContainer.cookie.secure : defaultContainer.cookie.secure
-                },
-                sessionPrefix: (newSessionContainer.sessionPrefix) ? newSessionContainer.sessionPrefix : defaultContainer.sessionPrefix,
-                genId: (newSessionContainer.genId) ? newSessionContainer.genId : defaultContainer.genId,
-                resave: (newSessionContainer.resave) ? newSessionContainer.resave : defaultContainer.resave,
-                rolling: (newSessionContainer.rolling) ? newSessionContainer.rolling : defaultContainer.rolling,
-                saveUninitialized: (newSessionContainer.saveUninitialized) ? newSessionContainer.saveUninitialized : defaultContainer.saveUninitialized,
-                store: (newSessionContainer.store) ? newSessionContainer.store : defaultContainer.store,
-            };
-        }
     };
 
     /**
